Moved searchLadokResults and createLadokResultObject to utils (they are basically the same
