+++ lib/transfer-examination.js
@@ -1,7 +1,13 @@
 /* eslint-disable no-await-in-loop,no-continue */
 const CanvasAPI = require("@kth/canvas-api");
 const log = require("skog");
-const { ladokGot, ladokSearch, getLadokGrade } = require("./utils");
+const {
+  ladokGot,
+  getLadokGrade,
+  createLadokResultObject,
+  searchLadokResults,
+  ResultObjectError,
+} = require("./utils");
 const mongo = require("./mongo");
 const {
   ladokGenericErrorHandler,
@@ -43,24 +49,6 @@ async function getCourseRounds(examinationRoundId) {
   return Array.from(new Set(allRounds));
 }
 
-async function searchLadokResults(examinationRoundId, mode) {
-  const courseRounds = await getCourseRounds(examinationRoundId);
-
-  if (courseRounds.length === 0) {
-    return [];
-  }
-
-  return ladokSearch(
-    `/resultat/studieresultat/rapportera/aktivitetstillfalle/${examinationRoundId}/sok`,
-    {
-      Filtrering: [mode === "create" ? "OBEHANDLADE" : "UTKAST"],
-      KurstillfallenUID: courseRounds,
-      LarosateID: process.env.LADOK_KTH_LAROSATE_ID,
-      OrderBy: ["EFTERNAMN_ASC", "FORNAMN_ASC", "PERSONNUMMER_ASC"],
-    }
-  );
-}
-
 /**
  *
  * @param courseId
@@ -84,11 +72,11 @@ async function getResults(courseId, assignmentId, token) {
 
   const examinationRounds = new Set(
     sections.map(getLadokId).filter((id) => id)
-  );
+  ); // diff
 
   log.info(
     `Course ${courseId} is mapped to ${examinationRounds.size} examination rounds`
-  );
+  ); // diff
 
   const submissions = await canvas
     .list(`/courses/${courseId}/assignments/${assignmentId}/submissions`, {
@@ -108,18 +96,30 @@ async function getResults(courseId, assignmentId, token) {
     mode: null,
   }));
 
-  for (const examination of examinationRounds) {
-    const results1 = await searchLadokResults(examination, "create");
-    const results2 = await searchLadokResults(examination, "update");
+  for (const examinationRoundId of examinationRounds) {
+    // diff
+    const courseRounds = await getCourseRounds(examinationRoundId); // diff
+
+    const resultsCreate = await searchLadokResults({
+      examinationRoundId, // diff
+      courseRounds,
+      mode: "create",
+    });
+    const resultsUpdate = await searchLadokResults({
+      examinationRoundId, // diff
+      courseRounds,
+      mode: "update",
+    });
 
     log.info(
-      `Found ${results1.length} results in Ladok for Create in examination ${examination}`
-    );
+      `Found ${resultsCreate.length} results in Ladok for Create in examination ${examinationRoundId}`
+    ); // diff
     log.info(
-      `Found ${results2.length} results in Ladok for Update in examination ${examination}`
-    );
+      `Found ${resultsUpdate.length} results in Ladok for Update in examination ${examinationRoundId}`
+    ); // diff
 
-    for (const result of results1) {
+    // ***** CREATE *****
+    for (const result of resultsCreate) {
       const submission = grades.find((s) => s.id === result.Student.Uid);
 
       if (!submission) {
@@ -137,12 +137,11 @@ async function getResults(courseId, assignmentId, token) {
           submission.grade
         ));
 
-      if (ladokGrade) {
-        submission.mode = "create";
-      }
+      submission.mode = ladokGrade && "create";
     }
 
-    for (const result of results2) {
+    // ***** UPDATE *****
+    for (const result of resultsUpdate) {
       const submission = grades.find((s) => s.id === result.Student.Uid);
 
       if (!submission) {
@@ -163,7 +162,7 @@ async function getResults(courseId, assignmentId, token) {
       const existingResult = result.ResultatPaUtbildningar.find(
         (rpu) =>
           rpu.Arbetsunderlag &&
-          rpu.Arbetsunderlag.AktivitetstillfalleUID === examination
+          rpu.Arbetsunderlag.AktivitetstillfalleUID === examinationRoundId // diff
       );
 
       if (
@@ -177,94 +176,6 @@ async function getResults(courseId, assignmentId, token) {
   }
 
   return grades;
-}
-
-class ResultObjectError extends Error {
-  constructor({ type, message }) {
-    super(message);
-    this.type = type;
-  }
-}
-
-/**
- * Create a result object that can be sent to LADOK.
- *
- * @param {string} param0.submissionType [create|update]
- * @param {array} param0.grades
- * @param {object} param0.result
- * @param {string} param0.moduleId
- * @param {string} param0.examinationDate
- * @returns LadokResultObject
- */
-async function createLadokResultObject({
-  actionType,
-  grades,
-  result,
-  utbildningsinstansUID,
-  examinationDate,
-}) {
-  const submission = grades.find((s) => s.id === result.Student.Uid);
-  if (!submission) {
-    // This should be logged with log.info
-    throw new ResultObjectError({
-      type: "missing_student",
-      message: `Missing student in Canvas ${result.Student.Uid} (${result.Student.Fornamn} ${result.Student.Efternamn} )`,
-    });
-  }
-
-  let ladokGrade;
-  if (submission?.grade) {
-    ladokGrade = await getLadokGrade(
-      result.Rapporteringskontext.BetygsskalaID,
-      submission.grade
-    );
-  }
-  if (!ladokGrade)
-    throw new ResultObjectError({
-      type: "missing_grade",
-      message: "No ladok grade found for this submission",
-    });
-
-  const commonData = {
-    Betygsgrad: ladokGrade.ID,
-    BetygsskalaID: result.Rapporteringskontext.BetygsskalaID,
-    Examinationsdatum: examinationDate,
-  };
-
-  if (actionType === "create") {
-    return {
-      ...commonData,
-      Uid: result.Uid, // TODO: Why is this different from the update object and same as below?
-      StudieresultatUID: result.Uid,
-      UtbildningsinstansUID: utbildningsinstansUID,
-    };
-  }
-
-  if (actionType === "update") {
-    const existingResult = result.ResultatPaUtbildningar.find(
-      (rpu) =>
-        rpu.Arbetsunderlag?.UtbildningsinstansUID === utbildningsinstansUID
-    );
-
-    // This submission already exists (NOOP)
-    if (existingResult.Arbetsunderlag.Betygsgrad === ladokGrade.ID)
-      throw new ResultObjectError({
-        type: "result_exists",
-        message: "This result already exists in LADOK",
-      });
-
-    return {
-      ...commonData,
-      Uid: result.Student.Uid,
-      ResultatUID: existingResult.Arbetsunderlag.Uid,
-      SenasteResultatandring:
-        existingResult.Arbetsunderlag.SenasteResultatandring,
-    };
-  }
-
-  throw new Error(
-    `Unknown actionType "${actionType}", should match [create|update]`
-  );
 }
 
 /**
@@ -321,11 +232,16 @@ async function transferResults(courseId, assignmentId, examinationDate, token) {
 
   // **** CREATE ****
   const resultsForLogCreated = [];
-  for (const examination of examinationRounds) {
-    const resultsCreate = await searchLadokResults(examination, "create");
+  for (const examinationRoundId of examinationRounds) {
+    const courseRounds = await getCourseRounds(examinationRoundId);
+    const resultsCreate = await searchLadokResults({
+      examinationRoundId,
+      courseRounds,
+      mode: "create",
+    });
 
     log.info(
-      `Found ${resultsCreate.length} results in Ladok for Create in examination ${examination}`
+      `Found ${resultsCreate.length} results in Ladok for Create in examination ${examinationRoundId}`
     );
 
     for (const result of resultsCreate) {
@@ -382,11 +298,16 @@ async function transferResults(courseId, assignmentId, examinationDate, token) {
 
   // **** UPDATE ****
   const resultsForLogUpdated = [];
-  for (const examination of examinationRounds) {
-    const resultsUpdate = await searchLadokResults(examination, "update");
+  for (const examinationRoundId of examinationRounds) {
+    const courseRounds = await getCourseRounds(examinationRoundId);
+    const resultsUpdate = await searchLadokResults({
+      examinationRoundId,
+      courseRounds,
+      mode: "update",
+    });
 
     log.info(
-      `Found ${resultsUpdate.length} results in Ladok for Update in examination ${examination}`
+      `Found ${resultsUpdate.length} results in Ladok for Update in examination ${examinationRoundId}`
     );
 
     for (const result of resultsUpdate) {
