+++ lib/transfer-module.js
@@ -7,25 +7,19 @@
 
 const CanvasAPI = require("@kth/canvas-api");
 const log = require("skog");
-const { ladokGot, ladokSearch, getLadokGrade } = require("./utils");
+const {
+  ladokGot,
+  getLadokGrade,
+  createLadokResultObject,
+  searchLadokResults,
+  ResultObjectError,
+} = require("./utils");
 const mongo = require("./mongo");
 const {
   ladokGenericErrorHandler,
   canvasGenericErrorHandler,
   LadokApiError,
 } = require("./errors");
-
-async function searchLadokResults(moduleId, courseRounds, mode) {
-  return ladokSearch(
-    `/resultat/studieresultat/rapportera/utbildningsinstans/${moduleId}/sok`,
-    {
-      Filtrering: [mode === "create" ? "OBEHANDLADE" : "UTKAST"],
-      KurstillfallenUID: courseRounds,
-      LarosateID: process.env.LADOK_KTH_LAROSATE_ID,
-      OrderBy: ["EFTERNAMN_ASC", "FORNAMN_ASC", "PERSONNUMMER_ASC"],
-    }
-  );
-}
 
 /**
  *
@@ -46,7 +40,7 @@ async function getResults(courseId, moduleId, assignmentId, token) {
     .toArray()
     .catch(canvasGenericErrorHandler);
 
-  log.info(`Course ${courseId} has ${sections.length} sections`);
+  log.info(`Course ${courseId} has ${sections.length} sections`); // diff
 
   const courseRounds = sections
     .filter((s) => s.integration_id)
@@ -70,13 +64,22 @@ async function getResults(courseId, moduleId, assignmentId, token) {
     mode: null,
   }));
 
-  const results1 = await searchLadokResults(moduleId, courseRounds, "create");
-  const results2 = await searchLadokResults(moduleId, courseRounds, "update");
+  const resultsCreate = await searchLadokResults({
+    moduleId, // diff
+    courseRounds,
+    mode: "create",
+  });
+  const resultsUpdate = await searchLadokResults({
+    moduleId, // diff
+    courseRounds,
+    mode: "update",
+  });
 
-  log.info(`Found ${results1.length} results in Ladok for Create`);
-  log.info(`Found ${results2.length} results in Ladok for Update`);
+  log.info(`Found ${resultsCreate.length} results in Ladok for Create`); // diff
+  log.info(`Found ${resultsUpdate.length} results in Ladok for Update`); // diff
 
-  for (const result of results1) {
+  // ***** CREATE *****
+  for (const result of resultsCreate) {
     const submission = grades.find((s) => s.id === result.Student.Uid);
 
     if (!submission) {
@@ -96,7 +99,8 @@ async function getResults(courseId, moduleId, assignmentId, token) {
     submission.mode = ladokGrade && "create";
   }
 
-  for (const result of results2) {
+  // ***** UPDATE *****
+  for (const result of resultsUpdate) {
     const submission = grades.find((s) => s.id === result.Student.Uid);
 
     if (!submission) {
@@ -117,7 +121,7 @@ async function getResults(courseId, moduleId, assignmentId, token) {
     const existingResult = result.ResultatPaUtbildningar.find(
       (rpu) =>
         rpu.Arbetsunderlag &&
-        rpu.Arbetsunderlag.UtbildningsinstansUID === moduleId
+        rpu.Arbetsunderlag.UtbildningsinstansUID === moduleId // diff
     );
 
     if (
@@ -130,93 +134,6 @@ async function getResults(courseId, moduleId, assignmentId, token) {
   }
 
   return grades;
-}
-
-class ResultObjectError extends Error {
-  constructor({ type, message }) {
-    super(message);
-    this.type = type;
-  }
-}
-
-/**
- * Create a result object that can be sent to LADOK.
- *
- * @param {string} param0.submissionType [create|update]
- * @param {array} param0.grades
- * @param {object} param0.result
- * @param {string} param0.moduleId
- * @param {string} param0.examinationDate
- * @returns LadokResultObject
- */
-async function createLadokResultObject({
-  actionType,
-  grades,
-  result,
-  moduleId,
-  examinationDate,
-}) {
-  const submission = grades.find((s) => s.id === result.Student.Uid);
-  if (!submission) {
-    // This should be logged with log.info
-    throw new ResultObjectError({
-      type: "missing_student",
-      message: `Missing student in Canvas ${result.Student.Uid} (${result.Student.Fornamn} ${result.Student.Efternamn} )`,
-    });
-  }
-
-  let ladokGrade;
-  if (submission?.grade) {
-    ladokGrade = await getLadokGrade(
-      result.Rapporteringskontext.BetygsskalaID,
-      submission.grade
-    );
-  }
-  if (!ladokGrade)
-    throw new ResultObjectError({
-      type: "missing_grade",
-      message: "No ladok grade found for this submission",
-    });
-
-  const commonData = {
-    Betygsgrad: ladokGrade.ID,
-    BetygsskalaID: result.Rapporteringskontext.BetygsskalaID,
-    Examinationsdatum: examinationDate,
-  };
-
-  if (actionType === "create") {
-    return {
-      ...commonData,
-      Uid: result.Uid, // TODO: Why is this different from the update object and same as below?
-      StudieresultatUID: result.Uid,
-      UtbildningsinstansUID: moduleId,
-    };
-  }
-
-  if (actionType === "update") {
-    const existingResult = result.ResultatPaUtbildningar.find(
-      (rpu) => rpu.Arbetsunderlag?.UtbildningsinstansUID === moduleId
-    );
-
-    // This submission already exists (NOOP)
-    if (existingResult.Arbetsunderlag.Betygsgrad === ladokGrade.ID)
-      throw new ResultObjectError({
-        type: "result_exists",
-        message: "This result already exists in LADOK",
-      });
-
-    return {
-      ...commonData,
-      Uid: result.Student.Uid,
-      ResultatUID: existingResult.Arbetsunderlag.Uid,
-      SenasteResultatandring:
-        existingResult.Arbetsunderlag.SenasteResultatandring,
-    };
-  }
-
-  throw new Error(
-    `Unknown actionType "${actionType}", should match [create|update]`
-  );
 }
 
 /**
@@ -295,7 +212,7 @@ async function transferResults(
 
   const commonParams = {
     grades,
-    moduleId,
+    utbildningsinstansUID: moduleId,
     examinationDate,
   };
