+++ lib/utils.js
@@ -85,10 +85,124 @@ async function ladokSearch(endpoint, criteria) {
   return result;
 }
 
+class ResultObjectError extends Error {
+  constructor({ type, message }) {
+    super(message);
+    this.type = type;
+  }
+}
+
+/**
+ * Create a result object that can be sent to LADOK.
+ *
+ * @param {string} param0.submissionType [create|update]
+ * @param {array} param0.grades
+ * @param {object} param0.result
+ * @param {string} param0.utbildningsinstansUID moduleId or examinationId
+ * @param {string} param0.examinationDate
+ * @returns LadokResultObject
+ */
+async function createLadokResultObject({
+  actionType,
+  grades,
+  result,
+  utbildningsinstansUID,
+  examinationDate,
+}) {
+  const submission = grades.find((s) => s.id === result.Student.Uid);
+  if (!submission) {
+    // This should be logged with log.info
+    throw new ResultObjectError({
+      type: "missing_student",
+      message: `Missing student in Canvas ${result.Student.Uid} (${result.Student.Fornamn} ${result.Student.Efternamn} )`,
+    });
+  }
+
+  let ladokGrade;
+  if (submission?.grade) {
+    ladokGrade = await getLadokGrade(
+      result.Rapporteringskontext.BetygsskalaID,
+      submission.grade
+    );
+  }
+  if (!ladokGrade)
+    throw new ResultObjectError({
+      type: "missing_grade",
+      message: "No ladok grade found for this submission",
+    });
+
+  const commonData = {
+    Betygsgrad: ladokGrade.ID,
+    BetygsskalaID: result.Rapporteringskontext.BetygsskalaID,
+    Examinationsdatum: examinationDate,
+  };
+
+  if (actionType === "create") {
+    return {
+      ...commonData,
+      Uid: result.Uid, // TODO: Why is this different from the update object and same as below?
+      StudieresultatUID: result.Uid,
+      UtbildningsinstansUID: utbildningsinstansUID,
+    };
+  }
+
+  if (actionType === "update") {
+    const existingResult = result.ResultatPaUtbildningar.find(
+      (rpu) =>
+        rpu.Arbetsunderlag?.UtbildningsinstansUID === utbildningsinstansUID
+    );
+
+    // This submission already exists (NOOP)
+    if (existingResult.Arbetsunderlag.Betygsgrad === ladokGrade.ID)
+      throw new ResultObjectError({
+        type: "result_exists",
+        message: "This result already exists in LADOK",
+      });
+
+    return {
+      ...commonData,
+      Uid: result.Student.Uid,
+      ResultatUID: existingResult.Arbetsunderlag.Uid,
+      SenasteResultatandring:
+        existingResult.Arbetsunderlag.SenasteResultatandring,
+    };
+  }
+
+  throw new Error(
+    `Unknown actionType "${actionType}", should match [create|update]`
+  );
+}
+
+async function searchLadokResults({
+  moduleId,
+  examinationRoundId,
+  courseRounds,
+  mode,
+}) {
+  if (courseRounds.length === 0) {
+    return [];
+  }
+
+  const ladokSearchPath =
+    examinationRoundId !== undefined
+      ? `/resultat/studieresultat/rapportera/aktivitetstillfalle/${examinationRoundId}/sok`
+      : `/resultat/studieresultat/rapportera/utbildningsinstans/${moduleId}/sok`;
+
+  return ladokSearch(ladokSearchPath, {
+    Filtrering: [mode === "create" ? "OBEHANDLADE" : "UTKAST"],
+    KurstillfallenUID: courseRounds,
+    LarosateID: process.env.LADOK_KTH_LAROSATE_ID,
+    OrderBy: ["EFTERNAMN_ASC", "FORNAMN_ASC", "PERSONNUMMER_ASC"],
+  });
+}
+
 module.exports = {
+  createLadokResultObject,
   ladokSearch,
   ladokGot,
   ladokGot2,
   getGradingScales,
   getLadokGrade,
+  searchLadokResults,
+  ResultObjectError,
 };
